Added support for a "master audio file"; in the absence of that, we now mix all the inputs to a WAV file with Replay Gain applied (fixed #4). Either one gets encoded as AAC audio in the output.
Also removed the Replay Gain analysis of the channels, and replaced it with an auto-scaling option. It may be better to apply an overall scaling to a whole soundtrack to maintain scale across tracks, the same applies to the Replay Gain on the audio track.
Also some work on the "Go" function to make it not use form state any more.